Refactor glance plugin to use API (images only)
Refactors glance plugin to use the glance API, and temporarily removes
the metadef plugin since it's somewhat more complex. The existing glance
plugins have been removed (since they required access to the internals
of glance). The new plugin is in a directory within plugins, and a
helper module added to allow access to python-glanceclient and
python-keystoneclient, along with the addition of configuration options
to allow the service to authenticate (both during initial indexing and
operation of the listener).