Document Listener and Manage Entrypoints Dev Docs
Per: https://review.openstack.org/#/c/196750/2

The entrypoints were updated:

Renames 'index' entrypoint (and cmd) to 'manage', and adds a command
'index' whose current only subcommand is 'sync' (with the intent that
we'll add more, like testing how in-sync things are).

Adds an entry point 'listener' for the notification listener, and
renames that cmd to 'listener'.

Change-Id: Ieda4274aa8122ef72fcf038779559cf3f3165aa4
Partial-Bug: #1471955