Change Glance notifications to be create_or_update
The glance image plugin uses two separate functions for create and update based
on the notification type. Under these functions it is called update on the
elasticsearch client.  This call will fail with a 404 error if the image
doesn't already exist.  The reason for it not existing may be numerous
(Create notifications dropped due to some sort of error, user deletes from
index, etc), but we can not worry about why it may not exist.  Instead this
should do create_or_update like functionality and if the image doesn't exist
on the update call, it should be inserted.

This patch makes use of the elastic search functionality called doc_as_upsert,
which natively provides the type of handling desired as stated above.

Change-Id: Icd97e15ed549b644d6b1aa9d019a8dd0b44defe7
Closes-Bug: #1479138