Enable configurable index names
The 'enabled plugin' interface now returns a resource type to plugin
mapping so that the internal details of indexing are less tied to the
API, and specifically so that Elasticsearch index names are configurable
without requiring clients to be aware of it.

The elasticsearch doc type is NOT expected to be configurable, for the
reason that there seems no compelling reason to allow it; for that
reason, 'type' should uniquely identify a plugin (and it's suggested
that be the openstack identifier for a resource). In the event that this
changes, 'resource_type' is currently a pseudonym for 'type' in the
search API btu could be expanded in plugin implementations to fully
abstract away the elasticsearch doc type.

Changes the glance plugin index name to 'searchlight'.

Change-Id: Ibf562df1c3cc6b0c17ee65886641cfa9e6468d5e
Implements: blueprint dont-expose-index-doctype