Fix Glance Indexing after Notification
After an update notification, Glance images do not seem
to be searchable. If you filter by the name of the updated
image it will not show up. However, if you re-index, it
will be searchable.

This is due to the code currently indexing a wrapper "doc"
object around the actual doc. This was inadvertantly left
in when switching from doc as upsert.

Change-Id: I782e04fea81ba4b9a52781fadd174d51a1aa7624
Closes-Bug: #1486781