Fix sorting behavior for common problematic fields
For some common fields (at this time, 'name') where we'd like the
field to be tokenized for searching, but would also like to sort the
values sensibly, there's no clear way to index the field as one type.
This patch mandates a not_analyzed subfield named 'raw' on any common
fields (regardless of how the name field itself is stored), and
intercepts sort terms on that field.

Change-Id: Ic7f76b887a7013c1126de87bc31f937c40c3b460
Closes-Bug: #1488236