Improve documentation for Liberty release
The documentation cross linking from the devstack rst to the
individual plugins is missing due to the devstack rst living
in a different directory. We just need to provide a bit better
linking from devstack as well as a bit more info on each plugin
for where to place to the configuration options for each plugin.

There is also opportunity to make the documentation between
plugins more consistent and provide release note information.

Change-Id: Ib64eeda8a9f5b59777135537de60f049952a4660
Closes-Bug: #1500647