Fix for Property mapping for Metadef properties fails
Property mapping for Metadef properties regressed to
"name" from "property".

Glance Metadef plugin at some point in previous commits
had a regression on the mapping for property name.
Previous it was "property" and now its changed to "name" which
causes the listener to fail on property updates with
objects and namespaces.

To recreate the bug. Try updating a property on a namespace
and check if the name of the property in the elastic search
document which will not have the updated name.

Change-Id: I1c4e8ec8f1b9648668c9957c88d913bd76834e9e
Closes-Bug: #1492579