Fix doc_count for nested fields
Aggregations for nested fields return a doc_count of matching
*nested* documents, which leads to results as described in the bug. This
patch adds a reverse_nested clause to the aggregation query, and
overwrites the doc_count with that of the reverse count (i.e. how
many parent documents match).

The patch also includes a refactoring of the faceting code and tests.
The previous implementation in plugins/base.py looked messy and didn't
lend itself to testing; the refactoring allows testing of the faceting
code separate from plugins, and means the plugin tests can focus on what
fields get faceted. There's a small refactoring of the functional tests
to tidy up some helper functions.

Change-Id: I258ae8a194a689cc5912d4aeaaf253c6c0b9db9a
Closes-Bug: #1503080