Index child plugins as part of parent
Runs indexing for child plugins as part of the parent;
errors if a child plugin is specified in the 'type' list to index sync.

When a parent is indexed, it will create mappings for any detected
child plugins, and index their data. Child plugins must implement
a property 'parent_plugin_type' whose value is the resource type of
the parent; the _parent mapping field is applied automatically.

This should assist with a restriction in elasticsearch 2.x whereby
parent mappings cannot exist prior to child mappings being created.
Ordering is such that the parent will come last. It is also necessary
in Elasticsearch 2.x to assign a _routing value when indexing child
documents; this is left to a separate patch.

Change-Id: I715b79a03491f102feee7d82fd1d49fb1cf29ae6
Closes-Bug: #1532972