Elasticsearch v2 client warning causes devstack to not stack
We added a warning message about the v2 ES client not working
with searchlight. This works fine, but a complete devstack
run wasn't tested. When I just stacked, it couldn't complete
because of using sys.exit(1) when the warning is emitted.
This changes it to (0) so that devstack will, well, err...
stack.

Once global requirements and searchlight requirements
are capped at v2, the exit code of 1 will be more appropriate.

Change-Id: I92bde9826ef789c390b4abc791cb384aef4d422b
Partial-Bug: #253177