Remove index API
The index API is a passthrough to elasticsearch. Given the changes
we're making with aliasing, and that we perform significant transforms
on raw data from the APIs, there a few legitimate reasons to do this.

A future suggested replacement is an API to request indexing of a
specific instance of a resource type that would go through the proper
serialization functions.

Change-Id: I070909176aace9c30638de7e31cd7545c8d84918
Implements: blueprint remove-index-api