Fix functional test [api] config section
https://review.openstack.org/#/c/253981/ moved the API config
options to their own section; this wasn't reflected in the functional
tests, which was only a problem because the shutdown code can't deal
with multiple worker processes.

Moved bind_host, bind_port and workers to an [api] section, and
hardcoded workers=0 (since the test runner is already spawning multiple
test processes, and multiple API processes within that seems
egregious).

Change-Id: I84aa45620f426025148365057cc9edf6bca6db6f
Closes-Bug: #1532306