Don't use _id path mapping
Elasticsearch 2.x doesn't allow _id field mapping (of the form
{"_id": {"path": "<field name>"}}), which was intended to allow
a document field to be treated as a mapping's id field. This behavior
has been disallowed in favor of explicitly providing the document id.
Since we don't need to do it (we're currently using it in the Metadef
mapping), we can stop this behavior before we move to es 2.x.

Change-Id: Ibd6b2a533c3c8eca835e8ca2e6e6e663a911a8a3
Closes-Bug: #1529009