Remove outdated pot files
These files are outdated and not needed, translation is done using our
translation server. The current scripts work even when these files do
not exist. Once there are translated files, the files will be imported
automatically again and updated.

Remove test for valid po/pot files, this is now done in the infra
scripts.

This is part of implementing the new infra setup for translations,
see spec
http://specs.openstack.org/openstack-infra/infra-specs/specs/translation_setup.html
for full details.

Change-Id: Ic320a7218840ab24cad34fa5b774e2107856d463