Moved CORS middleware configuration into oslo-config-generator
The default values needed for searchlight's implementation of cors
middleware have been moved from paste.ini into the configuration
hooks provided by oslo.config. Furthermore, these values have been
added to the default initialization procedure. This ensures
that if a value remains unset in the configuration file, it will
fallback to using sane defaults. It also ensures that an operator
modifying the configuration will be presented with that same
set of defaults.

Change-Id: Iada9cedcdc5b104bf2fa1c68d0d74794c04d1d28
Closes-Bug: 1551836