Remove pre-1.0 glance auth error handling
Prior to glance 1.0 (which didn't use keystone sessions) we implemented
custom authorization error retry code for glance. glanceclient 2.0.0
was no longer compatible with some of the tests for that code, and the
minimum requirement for glanceclient in mitaka is now 1.2.0. As a
result, this patch switches glance to using the same session-based auth
all the other clients do. This has been tested for gc 1.2 and 2.0.

Change-Id: Ida5b3c3bff5cc6e933a4c6dc0b7f5f429ddf4554
Closes-Bug: #1552875