Set doc_values on field definitions
Using the doc_values field storage format can result in dramatically
reduced memory usage at the cost of some disk space. This patch enforces
what is the default in Elasticsearch 2.x. It does so by modifying the
mapping to add doc_values: True for appropriate field types. This can
be overridden globally in configuration, per plugin via a property or
for individual field types.

See
https://www.elastic.co/guide/en/elasticsearch/guide/current/doc-values.html

This is an alternative to https://review.openstack.org/#/c/287401/

Change-Id: I5e5fb880657fcab7223e0e5f7525aec07fbdf612
Closes-Bug: #1498995