Fix _parent determination in some role sep configs
This patch attempts to make handling of the _parent field in some
parent/child role separation scenarios. The bulk action code generation
needs to be refactored but outside the scope of this patch which is to
fix some real behavioral problems, primarily where the parent or child
plugins (but not both) are under role separation. The action generation
code is not very consistent which makes this harder to grasp. This patch
adds several unit tests to test these conditions.

The situations that can occur, and how _parent is dealt with, are:

* parent is role separated
  * child is role separated: the _USER and _ADMIN docs each refer to a
        parent with the same suffix
  * child is NOT separated: this case is not encouraged, but where it
        exists the child documents will use the _USER parent document as
        their _parent to avoid security issues with has_parent query
        fishing
* parent is not role separated
  * child is role separated: both child docs will use the parent id
        without any prefix
  * child is not role separated: the simple case; no suffix in either
        case

Change-Id: I95df802079d530d5532e598ea66a684067c0e9e4
Closes-Bug: #1550440