Add _routing field to ES CRUD for plugins
Routing was optional until now, but with swift plugin it
has become mandatory as swift has multiple hierarchies of
parent and child relationships in document, wherein grand
child and grandparent have to be located in the same shard
for the queries to work.

This patchset adds _routing field while indexing and CRUD
operations on ES. By default _routing is left empty/none
which makes the existing plugins not effected by this change.
Plugins which require routing feature has to explicitly
specify in their plugin class.

Implements-blueprint: document-routing

Change-Id: I0264646e294fa6d87606cd034ba638d9fd03e5cb