Fix doc_values for dynamic_templates
Plugins with dynamic_templates were not getting correct mappings
created because of an error in the doc_values code. The code was
expecting dynamic_templates to contain a dict similar to the
'properties' in explicit mappings, where in fact it's a list with
somewhat different syntax. dynamic_templates should now have doc_values
applied correctly.

Change-Id: Iafea4d5cb0b6ad9cdc8a6adc9ab8acbb1c3d1470
Closes-Bug: #1557043