Remove resource_group_name as a per-plugin option
Because of significant difficulties with some query types
(particularly has_parent and has_child) that don't have obvious
solutions at this time, we're disabling the ability to specify
resource_group_name per plugin (and thus all plugins will coexist
in a single index). We hope to revert this once we have a good solution.

Change-Id: Iacb4c9f71c4019ff3ee9de81e30d6ffdaf63b7c2
Closes-Bug: #1558240