Fix duplicate entries displayed w/"index sync".
The "index sync" command will list the set of plugins to be
re-synced for the user. When we switched over to the Resource
Group Model, the code to display the plugins was based on the
older index model and implicitly relied on the child-parent
relationship to control what got displayed. We need to make this
relationship more explicit in determining the plugins to display.

Change-Id: I88af5e7a6837ea7529b487ee1f8fabeb2b783958
Closes-Bug: #1557303