Remove port.create.end handler from nova
Nova listens to port.create.end events with the aim of updating
instances to which ports may be attached. However, this creates
unnecessary churn because any pre-attached port.create event
 will be succeeded by the compute.create.end notification. The handler
itself is incorrect because it doesn't check device_owner. It's safest
to remove the handler for now until we have a better solution for
attaching/detaching interfaces.

Change-Id: I52cf0d6821e8acf54a3ad4397519c77d789f57dd
Closes-Bug: #1567525