Make func test indexing more consistent
Functional test indexing was previously a bit messy; some aspects of the
plugin indexing behavior was duplicated in the test code, not always in
a complete way. This patch modifies the test base class to call
plugin.initial_indexing which requires a little bit of modification to
how data is passed, but with the end result that indexing data as part
of a functional test gives an identical environment as it would in a
real deployment.

Where test code calls self._index, the 'docs' argument must be an
iterable of the same kind the plugin's 'get_objects' returns, which is
why for some plugins DictObj must be used.

Change-Id: I5a16dba4662eba60181e1604fc3eaf734eba851a
Closes-Bug: #1552919