Ensure consistency in mapping field types
Elasticsearch requires all field types in an index to be consistent
(and it's better for searching, too). ES-2.x actually enforces this
and won't allow mappings to be created. This patch fixes a couple of
fields and adds a unit test to check all plugins in the source tree.

Note that this does change the structure of the 'security_group' field
for nova servers to be a flat string (or array of strings). The other
option would be to rename the field on ports, or make it a nested type
on Ports also. There is a post-search filtering step that attempts to
return results as compatible with the APIs as possible, at least for
the meantime until we have a larger discussion about this problem.

Conflicts:
	searchlight/tests/unit/test_plugin.py

Change-Id: I7de3643bd5507743e01880c1885175faacef47b6
Closes-Bug: #1532010