Reduce number of API calls on nova boot
Replaces some of the nova API calls with updates. The code looks really
complicated (because it is) and the reason is that the notification
processing isn't sequential; unlike the unit test, the vast likelihood
in reality is that many of the notifications for an instance will be
received almost simultaneously. As such the partial updates request the
document from the index, and then check the existing state fields. The
update is submitted with the existing version as a kind of transaction
check. This seems to work ok, and most importantly the create.end event
should always add the actual current state.

Patch also adds support for pause, shutdown and suspend (and their
counterparts) such that there'll be a state change update followed by a
full API update on the .end event.

Unit tests include some examples of boot, delete and power change
events.

Change-Id: Ie2b089a2a2d0cd6bb45d0f11ab2a8c77ff1a8fb1
Implements: blueprint nova-optionally-disable-some-events