Image type should be a facet
Glance images can either be of type Image or of type Snapshot.
This is determined with the image_type property. This property
is ONLY set if the image is of type snapshot.

However, there is no facet for image_type which makes filtering
in the UI harder.

To make it consistent, Searchlight will insert a value('image')
in 'image_type' field for base images.
For images based on snapshot glance provides 'snapshot' as value

Change-Id: Ib4d73b12d82229a2bb85434f665d734b171557b7
Closes-Bug: #1569485