Enable notification messaging pools
Adds a 'notification_pool' option in the [listener] configuration
section. This is an alternative to adding a separate topic that will be
recommended for the Newton release following more testing; for now it is
an optional alternative.

To enable:

  [listener]
  notifications_pool = searchlight-listener

You will see a rabbitmq queue named searchlight-listener. Note that your
topic must match the service notifications_topic setting, whether it be
the default of 'notifications' or 'searchlight_indexer'.

Closes-Bug: #1583215
Change-Id: Ie40876e88059d0326efa0b4abaa76b8b64f0b86b