Return raw subfields on facets of analyzed string fields
Facet options don't work well on analyzed string fields right now
because the aggregation runs on the tokens. Many of the string
fields we might want to give facet options for, however, have
subfields used for sorting which are not analyzed. Facet should
returne these subfields instead of the actual field.

Change-Id: Iae0622186934d1583f18c6c3c480e936e418fc3e
Closes-Bug: #1558884