Don't accept domain-scoped tokens for auth
Domain scoped tokens are valid as far as keystone's concerned but it's
not clear how searchlight should deal with them (similar to many
openstack projects); all our resource types currently are project-based.
We may wish to revisit this at some point and perhaps deny access on a
per-plugin basis, but for now this resolves the error and removes the
ability for a domain admin to run an all_projects query.

Change-Id: I35f52a6c15992b8560c035c0d0eb033088cc51b9
Closes-Bug: #1579926