Expose 'version' as a search parameter
Elasticsearch supports 'version' as a boolean parameter in searches to
control whether or not a document's _version is returned. This patch
exposes it. Note that the value may not necessarily have meaning across
document types; some _versions will be a simple incrementing integer and
others will be somewhat time-based.

Change-Id: I54de37e63b906d353db103ff729c524b778a981a
Implements: blueprint expose-version