Enabled Neutron notifications in sample local.conf
For a period of time, neutron seemed to send notifications by default.
That period of time seems to have passed and notifications are not longer
on by default when you stack. At least that's what I observed on today's
devstack.  Whatever the case, it seems we might as well explicitly enable
them.

Change-Id: Ica4f526305e0a984c0d9bb689f50da42406dcf82