Log notification events
When running the listener in debug mode, there's a HUGE amount of
information output in logs (or the console). When running at INFO
there's no information logged at all. This patch adds a single line
that includes the document type, event type and metadata timestamp
as well as any other data that might be useful for a given type to
aid with problem diagnosis.

Change-Id: Ic5559d4564347dc684107c7bb2010733f22191df