Service policy file support
Adds support for configuring policy files for e.g. compute, image
etc. Each resource type defines a policy action it will look for.

Simplifies the existing searchlight policy.json to remove the
resource:operation rules in favor of allowing or disallowing access
to a resource for all operations. This patch also changes the default
rules for searchlight to admin_or_owner type rules which is more similar
to other services, and improves the plumbing to allow those kinds of
rules to work.

Change-Id: Ic705ee6ba864b646802c01a27e78a92db7273897
Implements: blueprint support-service-policy-files