Ignore nova deleting state changes
Treats the final two state transitions during server deletion
as a state change and ignored respectively. Adds more information
to the log message for nova servers, and adds a 'finished' log
message to assist with multithreaded debugging.

Marking this as partial because I think there is still an underlying
issue; the error state happens when the 'shutdown.end' event takes
longer to process than the 'delete.end' (which is always) AND some
confluence of timing happens such that the indexing operation for the
shutdown event manages to get past the conflict resolution.

Partial-Bug: #1588501
Change-Id: If6aebbe08d536c63be65f5cb6f550393077071ae