Add region name to mappings and documents
If resource_plugin.include_region_name is set, region_name will be
added to mappings and indexed documents. This enables the deployment
of multi-region searchlight though we will not be doing further work
on it in this release. Plugin configs can override the value in
service_credentials.os_region_name with 'override_region_name' for
cross-region shared services

Change-Id: I35f45e25227d9dff45b71bdefa48b4439ee7dfb4
Partially-implements: blueprint cross-region-searching