Warn user if types using es to reindex don't exist
If user specifies --type parameter in index syncing, this will
cause resource types that are explicit specified on the command
line are indexed in ElasticSearch through the associated plugin
making API calls. There is a possibility that old indices or
type mappings are not exist, reindexing these types from es
will result in incomplete index. Add check and
warning for users.

Co-Authored-By: Zhenyu Zheng <zhengzhenyu@huawei.com>

Change-Id: I8cd18aa0fa208a91e0b254ed2945923b4d9fa17a
Closes-Bug: #1596710