Add http_proxy_to_wsgi to api-paste
This sets up the HTTPProxyToWSGI middleware in front of Searchlight.
The purpose of thise middleware is to set up the request URL correctly
in case there is a proxy (For instance, a loadbalancer such as HAProxy)
in front of Searchlight.

So, for instance, when TLS connections are being terminated in the
proxy, and one tries to get the versions from the / resource of
Searchlight, one will notice that the protocol is incorrect; It will
show 'http' instead of 'https'. So this middleware handles such cases.

The HTTPProxyToWSGI is off by default and needs to be enabled via a
configuration value.

Change-Id: I79ef2f9340dd6b0c6eab8079fd5495f619d99adf
Closes-bug: #1590608
Co-Authored-By: abdul nizamuddin <abdul.nizamuddin@nectechnologies.in>