Add tools/test-setup.sh script to setup test environment
Add simple script to setup the test environment as needed by unit-tests.

This script can be used by users to setup their test environment.

It will be run by OpenStack CI systems as specific setup
before running unit tests.

This is needed on all active branches. This installs for liberty
elasticsearch 1 packages, this version was used for testing liberty when
it was released.

Also, switch to post-versioning, this is needed by newer pbr releases,
and fixes a type bug in one of the test cases that's been fixed on other
branches.

Co-Authored-By: Steve McLellan <steven.j.mclellan@gmail.com>
Depends-On: Ic6269e0a03c8540569f242ebc78c98a5922c8026
Change-Id: I56d8c5f21e5dc62e1bace93f79d8d017d6df07e7