Add alt text for badges
The RST image directive takes an alt paramter that's used to supply
an alt attribute for the HTML image element that's generated during
RST -> HTML conversion.  The alt text is useful for accessibility
purposes and is also displayed if the image source is unavailable
when the HTML is generated.  Because of the latter point, we can't
rely on the accessibility features of the SVG image, we need to
maintain some info here in the README.rst file.

Change-Id: I3d9e6710b6eadb01f2454d201b62b06b37794ab0