Replace oslo_utils.timeutils.isotime
Function 'oslo_utils.timeutils.isotime()' is deprecated in version '1.6'
and will be removed in a future version.
We cannot use 'datetime.datetime.isoformat()' instead,
because the format of the string generated by isoformat
isn't the same as the format of the string generated by isotime.
The string is used in tokens and other public APIs and we can't change it
without potentially breaking clients.

So the workaround is copying the current implementation from
oslo_utils.timeutils.isotime() to common/utils.py

For more informations:
http://docs.openstack.org/developer/oslo.utils/api/
timeutils.html#oslo_utils.timeutils.isotime

APIImpact

Closes-Bug: #1694352
Change-Id: I36aa771a2395ed35583c59c6563ea78e63cf6b8a