Implement basic policy module in code (1)
This change prepares the searchlight project to start implementing
policies in code. Subsequent patches will register more searchlight
policies in code and remove the corresponding entry from the policy
file maintained in source.
This is part of a community effort to provide better user experience
for those having to maintain RBAC policy.
More information on this effort can be found below:
https://governance.openstack.org/tc/goals/queens/policy-in-code.html

Change-Id: Ic3949a892df7b0a9ec2a8a189583bd72e0e031e8
Co-authored-By: Dai Dang-Van <daidv@vn.fujitsu.com>
Implements: blueprint policy-in-code