Add a vision-reflection document for Searchlight
The TC wrote a cloud vision [1] and then asked [2] projects to
do a self-evaluation related to that vision. And, since we have
just released a vision document of Searchlight [3] at Stein-2,
it is a good practice to produce an vision reflection document
to align our work with the rest of OpenStack. This patchset is
produced based on the discussion at [4].

[1] https://governance.openstack.org/tc/reference/technical-vision.html
[2] http://lists.openstack.org/pipermail/openstack-discuss/2019-January/001417.html
[3] https://docs.openstack.org/searchlight/latest/contributor/searchlight-vision.html
[4] https://etherpad.openstack.org/p/searchlight-tc-vision-self-eval

Change-Id: I093268758bd5c66ef1ba9d543e9d8ee8156c30f2
Signed-off-by: Trinh Nguyen <dangtrinhnt@gmail.com>