Retire the Searchlight project
As announced in openstack-discuss ML[1], Searchlight project
is retiring in Wallaby cycle.

This commit retires this repository as per process deinfed in
project-guide[2]. Anyone would like to maintain it again, please
revert back this commit and propose the re-adding of Searchlight
to governance.

The community wishes to express our thanks and appreciation to all of
those who have contributed to the Searchlight project over the years.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/764519
Needed-By: https://review.opendev.org/c/openstack/governance/+/764530

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018637.html
[2] https://docs.openstack.org/project-team-guide/repository.html#retiring-a-repository

Change-Id: I0c5c9a961cca19113f4d59e37cb16e3bc0ccf8e1